==> jupyter_notebooks/NB1_IdentifyProblem+DataClean.ipynb <==
# NB1- Identify The Problem & Data Cleaning

## Objectives

- Upload dataset to repository.
- Load data and save it under inputs/datasets/raw/data.csv
- Inspect the data and save it under inputs/datasets/data_clean_id/data.csv

## Inputs

Download  data set from Kaggle and import data.csv.

## Outputs

Generate Dataset:   inputs/datasets/raw/data.csv
                    inputs/datasets/data_clean_id/data.csv

## Additional Comments


---

# Change working directory

* We are assuming you will store the notebooks in a subfolder, therefore when running the notebook in the editor, you will need to change the working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os

current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Using Predictive Analysis To Predict Diagnosis of a Breast Tumour 

## 1. Identify the problem
Breast cancer is the most common malignancy among women, accounting for nearly 1 in 3 cancers diagnosed among women in the United States, and it is the second leading cause of cancer death among women. Breast Cancer occurs as a result of abnormal growth of cells in the breast tissue, commonly referred to as a tumour. A tumour does not mean cancer - tumours can be benign (not cancerous), pre-malignant (pre-cancerous), or malignant (cancerous). Tests such as MRI, mammogram, ultrasound and biopsy are commonly used to diagnose breast cancer performed.

### 1.1 Expected outcome
Given breast cancer results from breast fine needle aspiration (FNA) test (is a quick and simple procedure to perform, which removes some fluid or cells from a breast lesion or cyst (a lump, sore or swelling) with a fine needle similar to a blood sample needle). Since this build a model that can classify a breast cancer tumour using two training classification:
* 1= Malignant (Cancerous) - Present
* 0= Benign (Not Cancerous) -Absent

### 1.2 Objective 
Since the labels in the data are discrete, the predication falls into two categories, (i.e., Malignant or benign). In machine learning this is a classification problem. 
        
> *Thus, the goal is to classify whether the breast cancer is benign or malignant and predict the recurrence and non-recurrence of malignant cases after a certain period.  To achieve this, we have used machine learning classification methods to fit a function that can predict the discrete class of new input.*

Early-stage breast cancer often has a positive outlook with many people surviving due to treatment. Early-stage breast cancer can include several stages of breast cancer when cancer has not spread to distant areas of the body. Treatments can often cure or cause cancer to go into complete remission means that cancer cells are undetectable during testing. This project will help people get a quicker diagnosis and faster treatment helping to increase survival rates.

### 1.3 Identify data sources
* The dataset is on Kaggle it is publicly available, so it does not require a license to use, the data set contain 569 rows and 32 columns. Each row represents and person and the columns have different signs or textures of tumours. The dataset contains **569 samples of malignant and benign tumour cells**. 
* The first two columns in the dataset store the unique ID numbers of the samples and the corresponding diagnosis (M=malignant, B=benign), respectively. 
* The columns 3-32 contain 30 real-value features that have been computed from digitized images of the cell nuclei, which can be used to build a model to predict whether a tumour is benign or malignant. 

#### Getting Started: Load libraries and set options 

# load libraries
import numpy as np  # linear algebra
import pandas as pd  # data processing, CSV file I/O (e.g. pd.read_csv)

# Fetch raw data from Kaggle

First we need to install Kaggle to access the Kaggle API and to fetch the raw data.

! pip install kaggle==1.5.12

Make the Kaggle authentication token available for the session. 

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

KaggleDatasetPath = "vijayaadithyanvg/breast-cancer-prediction"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

* Define path to the Kaggle dataset we want to download 
* Indicate the destination folder for the downloaded data 
* Download the data

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

# Load and Inspect Kaggle data

* Load the dataset as a pandas Data Frame and assign it to df
* View the first five rows of the data in the df variable

data = pd.read_csv(f"inputs/datasets/raw/data.csv")
data.head()

#### Inspecting the data
The first step is to visually inspect the new data set. There are multiple ways to achieve this:
* The easiest being to request the first few records using the Data Frame data.head()* method. By default, “data.head()” returns the first 5 rows from the Data Frame object df (excluding the header row). 
* Alternatively, one can also use “df.tail()” to return the five rows of the data frame. 
* For both head and tail methods, there is an option to specify the number of records by including the required number in between the parentheses when calling either method. Inspecting the data

data.head(5)

You can check the number of cases, as well as the number of fields, using the shape method, as shown below.

# Id column is redundant and not useful, we want to drop it
data.drop("id", axis=1, inplace=True)
# data.drop('Unnamed: 0', axis=1, inplace=True)
data.head(2)

data.shape

In the result displayed, you can see the data has 569 records, each with 32 columns.

The **“info()”** method provides a concise summary of the data; from the output, it provides the type of data in each column, the number of non-null values in each column, and how much memory the data frame is using.

The method **get_dtype_counts()** will return the number of columns of each type in a DataFrame:

# Review data types with "info()".
data.info()

# Review number of columns of each data type in a Data Frame:
data.dtypes.value_counts()

From the above results, from the 32, variables, column id number 1 is an integer, diagnosis 569 non-null object. and rest are float. More on [python variables](https://www.tutorialspoint.com/python/python_variable_types.htm)

# check for missing variables
data.isnull().any()

data.diagnosis.unique()

From the results above, diagnosis is a categorical variable, because it represents a fix number of possible values (i.e., Malignant, of Benign. The machine learning algorithms wants numbers, and not strings, as their inputs so we need some method of coding to convert them.



---

# Push files to Repo

In case you don't need to push files to Repo, you may replace this section with "Conclusions and Next Steps" and state your conclusions and next steps.

# save the cleaner version of data frame for future analysis
data.to_csv("outputs/datasets/cleaned/data.csv")

> ### Now that we have a good intuitive sense of the data, Next step involves taking a closer look at attributes and data values. In notebook title :NB_NB2_ExploratoryDataAnalys, we will explore the data further
